fix(Event) Event class converted to Kotlin
