+++ bugsnag-android-core/src/test/java/com/bugsnag/android/EventApiTest.kt
@@ -30,39 +30,39 @@ internal class EventApiTest {
 
     @Test
     fun getUser() {
-        assertEquals(event.impl.userImpl, event.getUser())
+        assertEquals(event.userImpl, event.getUser())
     }
 
     @Test
     fun setUser() {
         event.setUser("99", "boo@example.com", "Boo")
-        assertEquals(User("99", "boo@example.com", "Boo"), event.impl.userImpl)
+        assertEquals(User("99", "boo@example.com", "Boo"), event.userImpl)
     }
 
     @Test
     fun addMetadataTopLevel() {
         event.addMetadata("foo", mapOf(Pair("wham", "bar")))
-        assertEquals(mapOf(Pair("wham", "bar")), event.impl.metadata.getMetadata("foo"))
+        assertEquals(mapOf(Pair("wham", "bar")), event.metadata.getMetadata("foo"))
     }
 
     @Test
     fun addMetadata() {
         event.addMetadata("foo", "wham", "bar")
-        assertEquals("bar", event.impl.metadata.getMetadata("foo", "wham"))
+        assertEquals("bar", event.metadata.getMetadata("foo", "wham"))
     }
 
     @Test
     fun clearMetadataTopLevel() {
         event.addMetadata("foo", mapOf(Pair("wham", "bar")))
         event.clearMetadata("foo")
-        assertNull(event.impl.metadata.getMetadata("foo"))
+        assertNull(event.metadata.getMetadata("foo"))
     }
 
     @Test
     fun clearMetadata() {
         event.addMetadata("foo", "wham", "bar")
         event.clearMetadata("foo", "wham")
-        assertNull(event.impl.metadata.getMetadata("foo", "wham"))
+        assertNull(event.metadata.getMetadata("foo", "wham"))
     }
 
     @Test
@@ -70,7 +70,7 @@ internal class EventApiTest {
         event.addFeatureFlag("demo_mode")
         assertEquals(
             listOf(FeatureFlag("demo_mode")),
-            event.impl.featureFlags.toList()
+            event.featureFlags.toList()
         )
     }
 
@@ -79,7 +79,7 @@ internal class EventApiTest {
         event.addFeatureFlag("sample_group", "a")
         assertEquals(
             listOf(FeatureFlag("sample_group", "a")),
-            event.impl.featureFlags.toList()
+            event.featureFlags.toList()
         )
     }
 
@@ -90,7 +90,7 @@ internal class EventApiTest {
         event.clearFeatureFlag("demo_group")
         assertEquals(
             listOf(FeatureFlag("sample_group", "a")),
-            event.impl.featureFlags.toList()
+            event.featureFlags.toList()
         )
     }
 
@@ -99,6 +99,6 @@ internal class EventApiTest {
         event.addFeatureFlag("demo_group")
         event.addFeatureFlag("sample_group", "a")
         event.clearFeatureFlags()
-        assertEquals(emptyList<FeatureFlag>(), event.impl.featureFlags.toList())
+        assertEquals(emptyList<FeatureFlag>(), event.featureFlags.toList())
     }
 }
