+++ bugsnag-plugin-react-native/src/test/java/com/bugsnag/android/TestHooks.java
@@ -2,7 +2,7 @@ package com.bugsnag.android;
 
 class TestHooks {
     static boolean getUnhandledOverridden(Event event) {
-        return event.getImpl().getUnhandledOverridden();
+        return event.getUnhandledOverridden();
     }
 
     static MetadataState generateMetadataState() {
